requirements.txt
pyotp
pyqrcode
pillow
tkinter
